add  delete validation, try-catch, authly by session, search by keyword, and other things. EP: Address, Career, Category ,  Person & Product

+++ pages/api/address/[id].ts
@@ -1,17 +1,24 @@
 import { Address } from '@prisma/client'
-import prisma from '../../../lib/prisma'
+import { withIronSessionApiRoute } from 'iron-session/next'
+import { ironOptions } from 'lib/ironSession'
+import prisma from 'lib/prisma'
 import type { NextApiRequest, NextApiResponse } from 'next'
+import { canUnserDo } from 'utils/checkPermissions'
 import z from 'zod'
 
-export default async function addressHandler(req: NextApiRequest, res: NextApiResponse) {
+export default withIronSessionApiRoute(addressHandler, ironOptions)
+
+async function addressHandler(req: NextApiRequest, res: NextApiResponse) {
   // Validate typeof id
   const idValidation = z.preprocess((value) => Number(value), z.number().positive())
 
   const {
     body,
     method,
-    query: { id }
+    query: { id },
+    session
   } = req
+  if (!canUnserDo(session, 'READ_ADDRESS')) return res.status(403).send(`Can't read this.`)
 
   const { success } = idValidation.safeParse(id)
   if (!success) return res.status(404).send(`Id ${id} Not Allowed`)
@@ -19,29 +26,53 @@ export default async function addressHandler(req: NextApiRequest, res: NextApiRe
   switch (method) {
     case 'GET':
       //obtenemos a UNA dirección
-      const address: Address | null = await prisma.address.findFirst({
-        where: { id: Number(id) }
-      })
-      if (!address) res.status(404).end(`Address not found`)
-      res.status(200).send(address)
+      try {
+        const address = await prisma.address.findFirst({
+          where: { id: Number(id) }
+        })
+        if (!address) res.status(404).end(`Address not found`)
+        res.status(200).send(address)
+      } catch (error) {
+        if (error instanceof Error) {
+          res.status(400).send(error.message)
+        }
+      }
       break
     case 'PUT':
+      if (!canUnserDo(session, 'EDIT_ADDRESS')) return res.status(403).send(`Can't edit this.`)
       //actualizamos a UNA dirección
-      const updateAddress: Address = await prisma.address.update({
-        data: {
-          ...body
-        },
-        where: {
-          id: Number(id)
+      try {
+        const address = await prisma.address.findFirst({
+          where: { id: Number(id) }
+        })
+        if (!address) res.status(404).end(`Address not found`)
+
+        const updateAddress = await prisma.address.update({
+          data: {
+            ...body
+          },
+          where: {
+            id: Number(id)
+          }
+        })
+        res.status(201).send(updateAddress)
+      } catch (error) {
+        if (error instanceof Error) {
+          res.status(400).send(error.message)
         }
-      })
-      if (!updateAddress) res.status(404).end(`Address not found`)
-      res.status(201).send(updateAddress || {})
+      }
       break
     case 'DELETE':
+      if (!canUnserDo(session, 'DELETE_ADDRESS')) return res.status(403).send(`Can't delete this.`)
       //eliminamos a UNA dirección
-      const delAddress: Address = await prisma.address.delete({ where: { id: Number(id) } })
-      res.status(202).send(delAddress)
+      try {
+        const delAddress: Address = await prisma.address.delete({ where: { id: Number(id) } })
+        res.status(202).send(delAddress)
+      } catch (error) {
+        if (error instanceof Error) {
+          res.status(400).send(error.message)
+        }
+      }
       break
     default:
       res.setHeader('Allow', ['GET', 'PUT', 'DELETE'])
